now locations are filtered to prevent duplicates
still needs an indicator for it tho

+++ app/Stores/GeocodeSlice.ts
@@ -3,6 +3,7 @@ import exp from "constants";
 import {json} from "stream/consumers";
 
 export type locationData = {
+    placeID: string
     address?: string | null,
     location: {
         lat: number | null,
@@ -12,6 +13,7 @@ export type locationData = {
 type GeocodeReturn = {
     results: [
         {
+            place_id: string
             formatted_address: string,
             geometry: {
                 location: {
@@ -31,9 +33,13 @@ const GeocodeSlice = createSlice({
         initialState,
         reducers: {
             setGeocodeData: (state: locationData[], action: PayloadAction<locationData>) => {
-                const newLocationsData = [action.payload, ...state]
-                localStorage.setItem("locations", JSON.stringify(newLocationsData))
-                return newLocationsData
+                const existingLocation  = state.find((location:locationData) => location.placeID === action.payload.placeID);
+                if (!existingLocation) {
+                    const newLocationsData = [action.payload, ...state];
+                    localStorage.setItem("locations", JSON.stringify(newLocationsData));
+                    return newLocationsData;
+                }
+                return state; // If the location already exists, return the current state
             },
             hydrateGeocodeData: (state: locationData[], action: PayloadAction<locationData[]>) => {
                 return action.payload
@@ -51,6 +57,7 @@ export const loadFromLocalStorage = () => {
 export const GeocodeCords = (lat: number, lng: number) => {
     return async (dispatch: Dispatch) => {
         const URL_Reverse = `https://maps.googleapis.com/maps/api/geocode/json?latlng=${lat},${lng}&result_type=administrative_area_level_2&key=${process.env.NEXT_PUBLIC_GOOGLE_MAPS_API_KEY}`
+        console.log(URL_Reverse)
         const fetchGeolocationData = async () => {
             const response = await fetch(URL_Reverse)
             if (!response.ok) return console.log("error fetching data")
@@ -59,6 +66,7 @@ export const GeocodeCords = (lat: number, lng: number) => {
         try {
             const locationData :GeocodeReturn = await fetchGeolocationData()
             dispatch(setGeocodeData({
+                placeID: locationData.results[0].place_id,
                 address:locationData.results[0].formatted_address,
                 location:locationData.results[0].geometry.location
             }))
